requirements.txt
numpy
pandas
scikit-learn
pyyaml
joblib
rich
pytest